[wip] start moving pretty-printing infrastructure to swarm-util
Generic pretty-printing infrastructure (`PrettyPrec` class, various
utility functions, etc.) can move to `-util` so we can use it
anywhere, for pretty-printing anything.  All the specific instances
are to be distributed so they live right next to the definitions of
the associated data types.
