+++ src/swarm-lang/Swarm/Language/Context.hs
@@ -33,6 +33,10 @@ instance ToJSON t => ToJSON (Ctx t) where
 instance FromJSON t => FromJSON (Ctx t) where
   parseJSON = genericParseJSON optionsUnwrapUnary
 
+instance (PrettyPrec t) => PrettyPrec (Ctx t) where
+  prettyPrec _ Empty = emptyDoc
+  prettyPrec _ (assocs -> bs) = brackets (hsep (punctuate "," (map prettyBinding bs)))
+
 -- | The semigroup operation for contexts is /right/-biased union.
 instance Semigroup (Ctx t) where
   (<>) = union
