+++ src/swarm-lang/Swarm/Language/Syntax.hs
@@ -73,7 +73,6 @@ module Swarm.Language.Syntax (
 
   -- * Terms
   Var,
-  DelayType (..),
   Term' (..),
   Term,
   TTerm,
