+++ swarm.cabal
@@ -453,7 +453,6 @@ library swarm-lang
     Swarm.Language.Parser.Util
     Swarm.Language.Pipeline
     Swarm.Language.Pipeline.QQ
-    Swarm.Language.Pretty
     Swarm.Language.Requirements
     Swarm.Language.Requirements.Analysis
     Swarm.Language.Requirements.Type
@@ -862,6 +861,7 @@ library swarm-util
     lens,
     minimorph,
     mtl,
+    prettyprinter,
     servant-docs,
     template-haskell,
     text,
@@ -877,6 +877,7 @@ library swarm-util
     Control.Carrier.Accum.FixedStrict
     Data.BoolExpr.Simplify
     Swarm.Language.Syntax.Direction
+    Swarm.Pretty
     Swarm.Util
     Swarm.Util.Effect
     Swarm.Util.Erasable
